make sure aarch64-neon-intrinsics.c test cases not duplicated
